+++ src/gcovr/formats/gcov/parser/common.py
@@ -45,7 +45,7 @@ class NegativeHits(Exception):
     ) -> None:
         """
         Raise exception if not ignored by options
-        >>> state = dict()
+        >>> state = dict(location=("file", 5))
         >>> NegativeHits.raise_if_not_ignored("code", None, state)
         Traceback (most recent call last):
             ...
@@ -77,7 +77,9 @@ class NegativeHits(Exception):
             if "negative_hits.warn_once_per_file" in persistent_states:
                 persistent_states["negative_hits.warn_once_per_file"] += 1
             else:
-                LOGGER.warning(f"Ignoring negative hits in line {line!r}.")
+                LOGGER.warning(
+                    f"Ignoring negative hits in {':'.join(str(item) for item in persistent_states['location'])}: {line!r}."
+                )
                 if "negative_hits.warn_once_per_file" in ignore_parse_errors:
                     persistent_states["negative_hits.warn_once_per_file"] = 1
         else:
@@ -103,7 +105,7 @@ class SuspiciousHits(Exception):
     ) -> None:
         """
         Raise exception if not ignored by options
-        >>> state = dict()
+        >>> state = dict(location=("file", 5))
         >>> SuspiciousHits.raise_if_not_ignored("code", None, state)
         Traceback (most recent call last):
             ...
@@ -135,7 +137,9 @@ class SuspiciousHits(Exception):
             if "suspicious_hits.warn_once_per_file" in persistent_states:
                 persistent_states["suspicious_hits.warn_once_per_file"] += 1
             else:
-                LOGGER.warning(f"Ignoring suspicious hits in line {line!r}.")
+                LOGGER.warning(
+                    f"Ignoring suspicious hits in {':'.join(str(item) for item in persistent_states['location'])}: {line!r}."
+                )
                 if "suspicious_hits.warn_once_per_file" in ignore_parse_errors:
                     persistent_states["suspicious_hits.warn_once_per_file"] = 1
         else:
@@ -151,11 +155,28 @@ def check_hits(
 ) -> int:
     """
     Check if hits count is negative or suspicious, if the issue is ignored returns 0
-    >>> check_hits(1, "", {}, 10, {})
+    >>> check_hits(1, "code", {}, 10, {"location": ("file", 5)})
     1
-    >>> check_hits(-1, "", {"all"}, 10, {})
+    >>> check_hits(-1, "code", {}, 10, {"location": ("file", 5)})
+    Traceback (most recent call last):
+        ...
+    gcovr.formats.gcov.parser.common.NegativeHits: Got negative hit value in gcov line 'code' caused by a
+    bug in gcov tool, see
+    https://gcc.gnu.org/bugzilla/show_bug.cgi?id=68080. Use option
+    --gcov-ignore-parse-errors with a value of negative_hits.warn,
+    or negative_hits.warn_once_per_file.
+    >>> check_hits(1000, "code", {}, 10, {"location": ("file", 5)})
+    Traceback (most recent call last):
+        ...
+    gcovr.formats.gcov.parser.common.SuspiciousHits: Got suspicious hit value in gcov line 'code' caused by a
+    bug in gcov tool, see
+    https://gcc.gnu.org/bugzilla/show_bug.cgi?id=68080. Use option
+    --gcov-ignore-parse-errors with a value of suspicious_hits.warn,
+    or suspicious_hits.warn_once_per_file or change the threshold
+    for the detection with option --gcov-suspicious-hits-threshold.
+    >>> check_hits(-1, "code", {"all"}, 10, {"location": ("file", 5)})
     0
-    >>> check_hits(1000, "", {"all"}, 10, {})
+    >>> check_hits(1000, "code", {"all"}, 10, {"location": ("file", 5)})
     0
     """
     if hits < 0: